Modify the way that the merging algorithm works
To make the merging algorithm scalable to more files, and make the merge
rules and mapping mechanisms more flexible, the algorithm was changed to
take a merge_tree and merge a doc_tree into it.  Thil way, any amount of
files can be merged, and any combination of changes from files mapped.